\name{limma_voom_DE-package}
\alias{limma_voom_DE-package}
\alias{limma_voom_DE}
\docType{package}
\title{
Differential Expression of count data using limma-voom
}
\description{
\packageDescription{limma_voom_DE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{limma_voom_DE}
\packageIndices{limma_voom_DE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{limma_voom_DE}

Maintainer: \packageMaintainer{limma_voom_DE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
